# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: CommerceCore API
  description: >
    Polyglot microservices e-commerce backend — catalog+cart-service (Node/TS)
    and order+payment-service (Spring Boot), fronted by an NGINX gateway.
    Month-1 MVP scope (see project roadmap Section 4).
  version: 0.1.0
servers:
  - url: http://localhost:8081
    description: Local gateway (Docker Compose)
...
paths:
...
  /auth/signup:
    post:
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already registered
  /auth/login:
    post:
      tags: [auth]
      summary: Authenticate and receive access + refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/refresh:
    post:
      tags: [auth]
      summary: Exchange a valid refresh token for a new access token (rotates refresh token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New token pair issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Refresh token invalid or expired

  # ---------------------------------------------------------------------
  # Products (Week 1, D6) — catalog only, no auth required to browse
  # ---------------------------------------------------------------------
...
components:
  schemas:
...
    SignupRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresInMs:
          type: integer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []