# services/order-payment-service/src/main/java/com/commercecore/orderpayment/dto/AuthResponse.java
package com.commercecore.orderpayment.dto;

public record AuthResponse(
        String accessToken,
        String refreshToken,
        String tokenType,
        long expiresInMs
) {
    public static AuthResponse of(String accessToken, String refreshToken, long expiresInMs) {
        return new AuthResponse(accessToken, refreshToken, "Bearer", expiresInMs);
    }
}