# services/order-payment-service/src/main/java/com/commercecore/orderpayment/dto/LoginRequest.java
package com.commercecore.orderpayment.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
        @NotBlank @Email String email,
        @NotBlank String password
) {}